requirements-runtime.txt
astunparse==1.6.3
autopep8==1.7.0
bs-highlighter==2.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.0.7
cycler==0.11.0
EditorConfig==0.12.4
fonttools==4.39.2
gitdb==4.0.9
GitPython==3.1.28
greenlet==2.0.2
idna==3.6
javalang==0.13.0
joblib==1.3.2
jsbeautifier==1.15.1
kiwisolver==1.4.4
lizard==1.17.10
matplotlib==3.7.1
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.3.5
pandasql==0.7.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pycg==0.0.5
pycodestyle==2.9.1
PyDriller==2.1
Pygments==2.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.4
regex==2023.12.25
requests==2.31.0
seaborn==0.12.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.45
termcolor==2.2.0
toml==0.10.2
torch==1.13.1
torchtext==0.5.0
tqdm==4.64.1
tree_sitter==0.20.1
types-pytz==2022.4.0.0
typing-extensions==3.7.4.3
urllib3==2.2.1
widlparser==1.0.11
